example/pyproject.toml
[project]
name = "example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataframetype",
    "pandera>=0.21.1",
]

[tool.uv.sources]
dataframetype = { path = "../" }

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pandas-stubs>=2.2.3.241126",
]
